+++ R/WordOfMouth-package.r
@@ -10,12 +10,11 @@
 #' }
 #' 
 #' \tabular{ll}{ Package: \tab WordOfMouth\cr Type: \tab Package\cr Version: 
-#' \tab 1.1.0\cr Date: \tab 2021-10-04\cr License: \tab GPL-3\cr Depends: \tab
+#' \tab 1.2.0\cr Date: \tab 2025-06-02\cr License: \tab GPL-3\cr Depends: \tab
 #' R (>= 3.0.1), methods\cr }
 #' 
 #' @name WordOfMouth-package
 #' @aliases WordOfMouth-package WordOfMouth
-#' @docType package
 #' @author Michael Scholz \email{michael.scholz@@th-deg.de}
 #' @author Thomas Woehner \email{Thomas.Woehner@@eah-jena.de}
 #' @author Ralf Peters \email{ralf.peters@@wiwi.uni-halle.de}
@@ -23,5 +22,5 @@
 #' @import methods LambertW
 #' @importFrom stats optim
 #' @keywords Word-of-Mouth 
-#' 
+#' "_PACKAGE"
 NULL