+++ NEWS.md
@@ -1,5 +1,11 @@
 # *News*
 
+# WordOfMouth 1.2.0 (2025-06-02)
+
+## Changes
+* New function for computing the effect of a WoM incentivization strategy
+* Improved readability of gradient function for optimizing the price
+
 # WordOfMouth 1.1.0 (2021-10-04)
 
 ## Changes